Consistency and refactor
