+++ streamlit/components/chain.py
@@ -1,89 +1,10 @@
 from ApiClient import ApiClient
 import streamlit as st
-
-
-@st.cache_data
-def cached_get_extensions():
-    return ApiClient.get_extensions()
-
-
-def prompt_selection(step_number, prompt):
-    available_prompts = ApiClient.get_prompts()
-    prompt_name = st.selectbox(
-        "Select Custom Prompt",
-        [""] + available_prompts,
-        index=available_prompts.index(prompt.get("prompt_name", "")) + 1
-        if "prompt_name" in prompt
-        else 0,
-        key=f"step_{step_number}_prompt_name",
-    )
-
-    if prompt_name:
-        prompt_args = ApiClient.get_prompt_args(prompt_name)
-        formatted_prompt_args = {
-            arg: st.text_input(
-                arg, value=prompt.get(arg, ""), key=f"{arg}_{step_number}"
-            )
-            for arg in prompt_args
-            if arg != "context" and arg != "command_list" and arg != "COMMANDS"
-        }
-        new_prompt = {
-            "prompt_name": prompt_name,
-            "prompt_args": formatted_prompt_args,
-        }
-        return new_prompt
-
-
-def command_selection(step_number, prompt):
-    agent_commands = cached_get_extensions()
-    available_commands = [cmd[0] for cmd in agent_commands]
-    command_name = st.selectbox(
-        "Select Command",
-        [""] + available_commands,
-        key=f"command_name_{step_number}",
-        index=available_commands.index(prompt.get("command_name", "")) + 1
-        if "command_name" in prompt
-        else 0,
-    )
-
-    if command_name:
-        command_args = ApiClient.get_command_args(command_name=command_name)
-        formatted_command_args = {
-            arg: st.text_input(
-                arg, value=prompt.get(arg, ""), key=f"{arg}_{step_number}"
-            )
-            for arg in command_args
-            if arg != "context" and arg != "command_list" and arg != "COMMANDS"
-        }
-        new_prompt = {
-            "command_name": command_name,
-            "command_args": formatted_command_args,
-        }
-        return new_prompt
-
-
-def chain_selection(step_number, prompt):
-    available_chains = ApiClient.get_chains()
-    chain_name = st.selectbox(
-        "Select Chain",
-        [""] + available_chains,
-        index=available_chains.index(prompt.get("chain_name", "")) + 1
-        if "chain_name" in prompt
-        else 0,
-        key=f"step_{step_number}_chain_name",
-    )
-    user_input = st.text_input(
-        "User Input",
-        value=prompt.get("user_input", ""),
-        key=f"user_input_{step_number}",
-    )
-
-    if chain_name:
-        new_prompt = {"chain_name": chain_name, "user_input": user_input}
-        return new_prompt
+from components.selectors import command_selection, prompt_selection, chain_selection
 
 
 def add_new_step(chain_name, step_number, agents):
+    st.markdown(f"## Add Chain Step {step_number}")
     agent_name = st.selectbox(
         "Select Agent",
         options=[""] + [agent["name"] for agent in agents],
@@ -97,11 +18,11 @@ def add_new_step(chain_name, step_number, agents):
     )
 
     if prompt_type == "Command":
-        prompt = command_selection(step_number=step_number, prompt={})
+        prompt = command_selection(step_number=step_number)
     elif prompt_type == "Prompt":
-        prompt = prompt_selection(step_number=step_number, prompt={})
+        prompt = prompt_selection(step_number=step_number)
     elif prompt_type == "Chain":
-        prompt = chain_selection(step_number=step_number, prompt={})
+        prompt = chain_selection(step_number=step_number)
     else:
         prompt = {}
     if st.button("Add New Step", key=f"add_step_{step_number}"):
@@ -146,11 +67,11 @@ def modify_step(chain_name, step, agents):
         key=f"prompt_type_{step_number}",
     )
     if prompt_type == "Command":
-        prompt = command_selection(step_number=step_number, prompt=step_prompt)
+        prompt = command_selection(prompt=step_prompt, step_number=step_number)
     elif prompt_type == "Prompt":
-        prompt = prompt_selection(step_number=step_number, prompt=step_prompt)
+        prompt = prompt_selection(prompt=step_prompt, step_number=step_number)
     elif prompt_type == "Chain":
-        prompt = chain_selection(step_number=step_number, prompt=step_prompt)
+        prompt = chain_selection(prompt=step_prompt, step_number=step_number)
     else:
         prompt = {}
     if st.button("Modify Step", key=f"modify_step_{step_number}"):
@@ -166,3 +87,30 @@ def modify_step(chain_name, step, agents):
             st.experimental_rerun()
         else:
             st.error("All fields are required.")
+
+
+def modify_chain(chain_name, agents):
+    if chain_name:
+        chain = ApiClient.get_chain(chain_name=chain_name)
+        st.markdown(f"## Modifying Chain: {chain_name}")
+        if chain:
+            for step in chain["steps"]:
+                if step is not None:
+                    new_step = modify_step(
+                        chain_name=chain_name,
+                        step=step,
+                        agents=agents,
+                    )
+
+        if len(chain["steps"]) > 0:
+            step_number = max([s["step"] for s in chain["steps"]]) + 1 if chain else 1
+        else:
+            step_number = 1
+
+        add_step = add_new_step(
+            chain_name=chain_name,
+            step_number=step_number,
+            agents=agents,
+        )
+    else:
+        st.warning("Please select a chain to manage steps.")
