+++ streamlit/pages/2-Chain_Management.py
@@ -2,10 +2,7 @@ import streamlit as st
 from ApiClient import ApiClient
 from components.verify_backend import verify_backend
 from components.docs import agixt_docs
-from components.chain import (
-    add_new_step,
-    modify_step,
-)
+from components.chain import modify_chain
 
 verify_backend()
 
@@ -24,50 +21,34 @@ chain_action = st.selectbox("Action", ["Create Chain", "Modify Chain", "Delete C
 
 if chain_action == "Create Chain":
     chain_name = st.text_input("Chain Name")
+    action_button = st.button("Create New Chain")
+    if action_button:
+        if chain_name:
+            ApiClient.add_chain(chain_name=chain_name)
+            st.success(f"Chain '{chain_name}' created.")
+            st.experimental_rerun()
+        else:
+            st.error("Chain name is required.")
+
+elif chain_action == "Delete Chain":
+    chain_name = st.selectbox("Chains", options=chain_names)
+    action_button = st.button("Delete Chain")
+    if action_button:
+        if chain_name:
+            ApiClient.delete_chain(chain_name=chain_name)
+            st.success(f"Chain '{chain_name}' deleted.")
+            st.experimental_rerun()
+        else:
+            st.error("Chain name is required.")
+
 elif chain_action == "Modify Chain":
-    chain_names = ApiClient.get_chains()
     selected_chain_name = st.selectbox("Select Chain", [""] + chain_names)
     if selected_chain_name:
-        chain = ApiClient.get_chain(chain_name=selected_chain_name)
-        st.markdown(f"## Modifying Chain: {selected_chain_name}")
-        if chain:
-            for step in chain["steps"]:
-                if step is not None:
-                    new_step = modify_step(
-                        chain_name=selected_chain_name,
-                        step=step,
-                        agents=agents,
-                    )
-
-        if len(chain["steps"]) > 0:
-            step_number = max([s["step"] for s in chain["steps"]]) + 1 if chain else 1
-        else:
-            step_number = 1
-
-        st.markdown(f"## Add New Step {step_number}")
-        add_step = add_new_step(
-            chain_name=selected_chain_name,
-            step_number=step_number,
-            agents=agents,
-        )
+        chain = modify_chain(chain_name=selected_chain_name, agents=agents)
     else:
         st.warning("Please select a chain to manage steps.")
 else:
-    chain_name = st.selectbox("Chains", ApiClient.get_chains())
-if chain_action == "Create Chain" or chain_action == "Delete Chain":
-    action_button = st.button("Perform Action")
-    if action_button:
-        if chain_name:
-            if chain_action == "Create Chain":
-                ApiClient.add_chain(chain_name=chain_name)
-                st.success(f"Chain '{chain_name}' created.")
-                st.experimental_rerun()
-            elif chain_action == "Delete Chain":
-                ApiClient.delete_chain(chain_name=chain_name)
-                st.success(f"Chain '{chain_name}' deleted.")
-                st.experimental_rerun()
-        else:
-            st.error("Chain name is required.")
+    chain_name = st.selectbox("Chains", options=chain_names)
 
 st.markdown("### Predefined Injection Variables")
 st.markdown(
